+++ deeper/levels/basic_level.py
@@ -6,9 +6,11 @@ from deeper import Block
 from deeper.components.sprite_vu import SpriteVu
 from deeper.catalog import Catalog
 
-class TestLevel(Level):
-    def __init__(self, timed=False):
+class BasicLevel(Level):
+    def __init__(self, rows=32, cols=32, timed=False):
         super().__init__('test', timed)
+        self.rows = rows
+        self.cols = cols
         self.create_grid()
         self.create_layer('Floor')
         self.create_layer('Default')
@@ -18,8 +20,8 @@ class TestLevel(Level):
         blueprint = catalog.find('Cell')
         size = glm.vec3(CELL_WIDTH, 0.01, 1)
         layer = self.create_layer('Grid')
-        for ty in range(0, 16):
-            for tx in range(0, 16):
+        for ty in range(0, self.rows):
+            for tx in range(0, self.cols):
                 position = glm.vec3(tx * CELL_WIDTH, 0, ty)
                 # position = glm.vec3(tx * CELL_WIDTH - CELL_HALF_WIDTH, 0, ty - CELL_HALF_DEPTH)
                 # print("position: ", position)