+++ deeper/model.py
@@ -1,6 +1,6 @@
 from sqlalchemy.orm import registry
 from sqlalchemy.orm.decl_api import DeclarativeMeta
-from sqlalchemy.ext.declarative import as_declarative, declared_attr
+from sqlalchemy.ext.declarative import declared_attr
 
 mapper_registry = registry()
 
@@ -13,6 +13,7 @@ class Model(metaclass=DeclarativeMeta):
 
     # these are supplied by the sqlalchemy2-stubs, so may be omitted
     # when they are installed
+    
     registry = mapper_registry
     metadata = mapper_registry.metadata