+++ deeper/views/scene_editor.py
@@ -23,8 +23,16 @@ class SceneEditor(SceneView):
             resolve_resource_path(":deeper:fonts/Roboto-Regular.ttf"), 16
         )
 
-    def save(self):
-        self.world.save(resolve_resource_path(":deeper:levels/"))
+    def new(self):
+        clock.schedule_once(lambda dt, *args, **kwargs : self._new(), 0)
+
+    def _new(self):
+        from .level_editor import LevelEditor
+        from ..state import WorldEditState
+        from ..levels.basic_level import BasicLevel
+        level = BasicLevel()
+        view = LevelEditor(self.window, WorldEditState(level))
+        self.window.show_view(view)
 
     def load(self):
         clock.schedule_once(lambda dt, *args, **kwargs : self._load(), 0)
@@ -36,13 +44,17 @@ class SceneEditor(SceneView):
         view = LevelEditor(self.window, WorldEditState(level))
         self.window.show_view(view)
 
+    def save(self):
+        self.world.save(resolve_resource_path(":deeper:levels/"))
+
     def create_menubar(self, children):
         children = [
             Menu(
                 "File",
                 [
-                    MenuItem("Save", lambda: self.save()),
+                    MenuItem("New", lambda: self.new()),
                     MenuItem("Load", lambda: self.load()),
+                    MenuItem("Save", lambda: self.save()),
                     MenuItem("Exit", lambda: exit(1)),
                 ],
             ),
@@ -73,6 +85,8 @@ class SceneEditor(SceneView):
         return menu
 
     def close_window(self, title):
+        if not title in self.windows:
+            return
         window = self.windows[title]
         self.windows.pop(title, None)
         self.gui.remove_child(window)