+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# DeeperRPG release 0.2.0 (2023-03-12)
+
+## Features
+
+Heirachical entities
+Basic level load and save
+
 # DeeperRPG release 0.1.9 (2023-02-28)
 
 ## Bug fixes